requirements.txt
PyQt6>=6.5.0
mss>=9.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.24.0
opencv-python-headless>=4.8.0
deep-translator>=1.11.0
google-generativeai>=0.5.0
markdown>=3.5.0
